Revert "Allow drag and drop of dimensions and measures within a query. (#212)"
This reverts commit 220f34e55fe33b6a3608b187d3bf30d1a8a86309.
