+++ package.json
@@ -51,9 +51,6 @@
   },
   "homepage": "https://github.com/malloydata/malloy-explorer#readme",
   "dependencies": {
-    "@dnd-kit/core": "^6.3.1",
-    "@dnd-kit/sortable": "^10.0.0",
-    "@dnd-kit/utilities": "^3.2.2",
     "@floating-ui/react-dom": "2.1.2",
     "@radix-ui/react-dialog": "^1.1.6",
     "@radix-ui/react-dropdown-menu": "^2.1.6",
